We no longer try to serialize an Exception object when one is raised in a sandbox.
Instead we store the important attributes of an exception object(#message, #class, #backtrace) in
a Hash object, serialize the hash object, and convert it to an OpenStruct in the parent process.

All exception attributes are sent back as strings, and the return value of #message, #class,
and #backtrace are evaluated in the context of the sandbox to avoid a user storiing mailicious code
in #message for example, which is executed outside of the sandbox when read from the parent.