Improve the test suite for Dia::SharedFeatures#exit_status.
I applied some general cleaning to the suite regarding #exit_status, and also
applied a baretest tag to the suite.

As of now, selectors(or tags) seem to be broken on baretest 0.5.0.pre but will come
in useful in the future.

I've also refactored the suite to take advantage of exercise/verify, something I haven't been
doing but I'm changing that(slowly).