Process.wait2 replaces Process.detach in #terminate
This means that #terminate will be a blocking method, but in reality, it
shouldn't block for more than one second(if at all).

Process.wait2 is only called in the event that Process.kill leaves a zombie
process in the process table, and it will(in my tests) reap the exit status
immediately.