Assert that Dia can handle an exception object which causes an exception when serializing data.
This feature has a slight flaw, consider the following:

      sandbox = Dia::RubyBlock.new(Dia::Profiles::NO_OS_SERVICES) do
        e = Exception.new
        def e.message
          raise()
        end
        raise(e)
      end
      sandbox.rescue_exception = true
      sandbox.run

The second exception(raise without arguments) will re-raise the Exception object referenced by
"e", which will cause another exception when Dia tries to access Exception#message again.

This exception goes uncaught - it is a scenario I'd like to find a decent solution to in the
future.

FYI, this scenario presents itself because Ruby will raise the last-raised exception(stored in $!)
if raise is given no arguments and $! doesn't reference nil. Additionally, $! is read-only(at least
on 19)